Consolidate selectors list to a single file and add template files
This commit introduces quite a big change to how the extensions are
developed. Prior to this commit the CSS selectors list was duplicated
across multiple files in each browser extension, and making a change to
the selectors required manually editing every single file where they were
duplicated. With the changes in this commit the selectors are now
consolidated in a single file (tools/selectors.txt) and .template files
have been introduced for each file that uses the selectors. To make a
change to the selectors that are used by the extensions, simply update
selectors.txt and invoke tools/process_templates.sh to propagate the
changes.

Hopefully this will make it slightly less frustrating to make changes to
the selectors in the future.