Don’t join selectors with newlines when processing templates
Using newlines between every selector in the processed template output
limits the places where we can use the selector template variables (e.g.
it can’t be used in a JavaScript string literal).
Also, since the .template files are the files that are edited, the
readability of the output files isn’t very important either.