[Safari] Convert Safari App Extension to Safari Extension
The Safari _App_ Extension support was introduced at WWDC 2016 and is
what I’ve been using up until now. However, it never really felt suitable
for NoSummaryForYou since the entire product is the extension, i.e. it’s
not something that is a companion to a native app.

This commit converts the Safari App Extension to the “old” Safari
Extension format which doesn’t include any native code. It also means
that the Safari extension won’t have to be distributed through the Mac
App Store, which I see as a win. Instead, the extension can just be
downloaded from the website or from the Safari Extension Gallery (if we
submit it there), similar to the Chrome and Firefox extensions IIRC.